wip medical folder functions

+++ src/main/java/epicode/it/healthdesk/entities/address/AddressSvc.java
@@ -4,6 +4,11 @@ import epicode.it.healthdesk.entities.address.city.City;
 import epicode.it.healthdesk.entities.address.city.CitySvc;
 import epicode.it.healthdesk.entities.address.dto.AddressMapper;
 import epicode.it.healthdesk.entities.address.dto.AddressRequest;
+import epicode.it.healthdesk.entities.appointment.Appointment;
+import epicode.it.healthdesk.entities.appointment.AppointmentRepo;
+import epicode.it.healthdesk.entities.appointment.AppointmentStatus;
+import epicode.it.healthdesk.entities.appointment.AppointmentSvc;
+import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.BeanUtils;
@@ -18,6 +23,7 @@ import java.util.List;
 public class AddressSvc {
     private final AddressRepo addressRepo;
     private final CitySvc citySvc;
+    private final AppointmentRepo appointmentRepo;
     private final AddressMapper mapper;
 
     public Address create(@Valid AddressRequest request) {
@@ -34,8 +40,19 @@ public class AddressSvc {
         return addressRepo.findById(id).orElse(null);
     }
 
+    @Transactional
     public void delete(Long id) {
         Address a = getById(id);
+        List<Appointment> appointments = appointmentRepo.findByDoctorAddress(a).stream().filter(a1 -> a1.getStatus() != AppointmentStatus.CANCELLED).toList();
+        ;
+        if (appointments.size() > 0)
+            throw new IllegalArgumentException("Ci sono appuntamenti prenotati a questo indirizzo, impossibile cancellare");
+
+        List<Appointment> cancelled = appointmentRepo.findByDoctorAddress(a).stream().filter(a1 -> a1.getStatus() == AppointmentStatus.CANCELLED).toList();
+
+        if (cancelled.size() > 0)
+            cancelled.forEach(app -> app.setDoctorAddress(null));
+
         addressRepo.delete(a);
     }
